Implement google oauth

+++ app.py
@@ -8,9 +8,19 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from database_setup import Restaurant, Base, MenuItem
 import json
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+from flask import make_response
+import requests
 
 
 app = Flask(__name__)
+
+
+CLIENT_ID = json.loads(
+    open('client_secret.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Restaurant Menu App"
 
 
 # Create db session
@@ -29,7 +39,119 @@ def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.
                     digits) for x in range(32))
     login_session['state'] = state
-    return 'The current session state is {}'.format(login_session['state'])
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(
+            json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorisation code
+    code = request.data
+
+    try:
+        # Upgrade the authorisation code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorisation code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token={}'
+           .format(access_token))
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1].decode())
+    # Abort if there is an error in the access token
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(
+            json.dumps('Current user is already connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h3>Welcome, '
+    output += login_session['email']
+    output += '!</h3>'
+    flash('You are now logged in as {}'.format(login_session['email']))
+    return output
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+@app.route('/logout/')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        print('Access Token is None')
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    print('In gdisconnect access token is {}'.format(access_token))
+    print('User name is: ')
+    print(login_session['email'])
+    url = 'https://accounts.google.com/o/oauth2/revoke?token={}'.format(
+        login_session['access_token'])
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print('result is ')
+    print(result)
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['email']
+        response = make_response(json.dumps(
+            'Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
 
 
 @app.route('/')
@@ -73,6 +195,9 @@ def restaurantMenuAPI(restaurant_id):
 @app.route('/restaurants/<int:restaurant_id>/new/',
            methods=['GET', 'POST'])
 def newMenuItem(restaurant_id):
+    # Restrict new menu item page to logged in users
+    if 'email' not in login_session:
+        return redirect('/login/')
     if request.method == 'POST':
         new_item = MenuItem(name=request.form['name'],
                             description=request.form['description'],
@@ -91,6 +216,9 @@ def newMenuItem(restaurant_id):
 @app.route('/restaurants/<int:restaurant_id>/edit/<int:menu_id>/',
            methods=['GET', 'POST'])
 def editMenuItem(restaurant_id, menu_id):
+    # Restrict edit menu item page to logged in users
+    if 'email' not in login_session:
+        return redirect('/login/')
     edit_item = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         if request.form['name']:
@@ -110,6 +238,9 @@ def editMenuItem(restaurant_id, menu_id):
 @app.route('/restaurants/<int:restaurant_id>/delete/<int:menu_id>/',
            methods=['GET', 'POST'])
 def deleteMenuItem(restaurant_id, menu_id):
+    # Restrict delete menu item page to logged in users
+    if 'email' not in login_session:
+        return redirect('/login/')
     delete_item = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         session.delete(delete_item)
